refactor directory layout
