+++ README.md
@@ -1,1 +1,3 @@
-PinningServiceAPI
+# PinningServiceAPI
+
+The `generated` directory is where the default code generation resides. It is also the directory that the command `make gen` stores data in. The `api` directory is where all of the API logic is contained that leverages the generated code.
