+++ api/routers.go
@@ -1,28 +1,29 @@
 /*
- * IPFS Pinning Service
+ * IPFS Pinning Service API
  *
- * This is the IPFS Pinning Service API. It attempts to be an implementation-agnostic API.
+ * This is the IPFS Pinning Service API. It attempts to be an implementation-agnostic API designed to be implemented by Pinning Service Providers.  This specification does **NOT** define the IPFS pinning API.  GitHub discussion around this API can be found in [ipfs/notes](https://github.com/ipfs/notes/issues/378).   **Critical Notice**  To allow backend systems behind the API to associate pins with a particular user, the API requires that a JSON Web Token is passed with requests that contains a `sub` claim (as specified by [OpenIDConnect Standard Claims](https://openid.net/specs/openid-connect-core-1_0.html#StandardClaims)) that represents the user's identity within the backend system, so as to associate the user with their pins. This allows the backend system to use the identity as a form of reference counting on pins, therefore only removing pins once no other users have it pinned.
  *
- * API version: v0.0.0
+ * API version: 0.0.1
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package api
 
 import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	"github.com/RTradeLtd/PinningServiceAPI/api/models"
 )
 
 // Route is the information for every URI.
 type Route struct {
 	// Name is the name of this Route.
-	Name        string
+	Name string
 	// Method is the string for the HTTP method. ex) GET, POST etc..
-	Method      string
+	Method string
 	// Pattern is the pattern of the URI.
-	Pattern     string
+	Pattern string
 	// HandlerFunc is the handler function of this route.
 	HandlerFunc gin.HandlerFunc
 }
@@ -58,42 +59,42 @@ var routes = Routes{
 	{
 		"Index",
 		http.MethodGet,
-		"/api/v0/",
+		"/",
 		Index,
 	},
 
 	{
 		"PinsCidDelete",
 		http.MethodDelete,
-		"/api/v0/pins/:cid",
-		PinsCidDelete,
+		"/pins/:cid",
+		models.PinsCidDelete,
 	},
 
 	{
 		"PinsCidGet",
 		http.MethodGet,
-		"/api/v0/pins/:cid",
-		PinsCidGet,
+		"/pins/:cid",
+		models.PinsCidGet,
 	},
 
 	{
 		"PinsCidPost",
 		http.MethodPost,
-		"/api/v0/pins/:cid",
-		PinsCidPost,
+		"/pins/:cid",
+		models.PinsCidPost,
 	},
 
 	{
 		"PinsGet",
 		http.MethodGet,
-		"/api/v0/pins",
-		PinsGet,
+		"/pins",
+		models.PinsGet,
 	},
 
 	{
 		"PinsPost",
 		http.MethodPost,
-		"/api/v0/pins",
-		PinsPost,
+		"/pins",
+		models.PinsPost,
 	},
 }
