--- go/model_error.go
@@ -1,17 +1,0 @@
-/*
- * IPFS Pinning Service
- *
- * This is the IPFS Pinning Service API. It attempts to be an implementation-agnostic API.
- *
- * API version: v0.0.0
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-type Error struct {
-
-	Code int32 `json:"code,omitempty"`
-
-	Message string `json:"message,omitempty"`
-}
