--- go/model_pin.go
@@ -1,19 +1,0 @@
-/*
- * IPFS Pinning Service
- *
- * This is the IPFS Pinning Service API. It attempts to be an implementation-agnostic API.
- *
- * API version: v0.0.0
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-type Pin struct {
-
-	Cid Cid `json:"cid,omitempty"`
-
-	Metadata map[string]string `json:"metadata,omitempty"`
-
-	Replication int32 `json:"replication,omitempty"`
-}
