--- go/model_pin_status.go
@@ -1,25 +1,0 @@
-/*
- * IPFS Pinning Service
- *
- * This is the IPFS Pinning Service API. It attempts to be an implementation-agnostic API.
- *
- * API version: v0.0.0
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-type PinStatus struct {
-
-	Peername string `json:"peername,omitempty"`
-
-	Peer string `json:"peer,omitempty"`
-
-	Error string `json:"error,omitempty"`
-
-	Pin Pin `json:"pin,omitempty"`
-
-	Status string `json:"status,omitempty"`
-
-	Timestamp string `json:"timestamp,omitempty"`
-}
