--- main.go
@@ -1,31 +1,0 @@
-/*
- * IPFS Pinning Service
- *
- * This is the IPFS Pinning Service API. It attempts to be an implementation-agnostic API.
- *
- * API version: v0.0.0
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package main
-
-import (
-	"log"
-
-	// WARNING!
-	// Change this to a fully-qualified import path
-	// once you place this file into your project.
-	// For example,
-	//
-	//    sw "github.com/myname/myrepo/go"
-	//
-	sw "./go"
-)
-
-func main() {
-	log.Printf("Server started")
-
-	router := sw.NewRouter()
-
-	log.Fatal(router.Run(":8080"))
-}
