+++ openapi_spec.yml
@@ -1,11 +1,35 @@
 openapi: 3.0.0
 info:
-  version: 'v0.0.0'
-  title: 'IPFS Pinning Service'
-  description: 'This is the IPFS Pinning Service API. It attempts to be an implementation-agnostic API.'
+  version: "0.0.1"
+  title: 'IPFS Pinning Service API'
+  description: "This is the IPFS Pinning Service API. It attempts to be an 
+    implementation-agnostic API designed to be implemented by Pinning Service 
+    Providers. 
+    
+    
+    This specification does **NOT** define the IPFS pinning API.
+    
+    
+    GitHub discussion around this API can be found in [ipfs/notes](https://github.com/ipfs/notes/issues/378).
+
+
+
+    **Critical Notice**
+    
+    
+    To allow backend systems behind the API to associate pins with a particular 
+    user, the API requires that a JSON Web Token is passed with requests that 
+    contains a `sub` claim (as specified by 
+    [OpenIDConnect Standard Claims](https://openid.net/specs/openid-connect-core-1_0.html#StandardClaims))
+    that represents the user's identity within the backend system, so as to 
+    associate the user with their pins. This allows the backend system to use 
+    the identity as a form of reference counting on pins, therefore only 
+    removing pins once no other users have it pinned."
+    
 servers:
+  - url: https://api.example.com
   - description: SwaggerHub API Auto Mocking
-    url: https://virtserver.swaggerhub.com/api/v0
+    url: https://virtserver.swaggerhub.com/lanzafame/ipfs-pinning-service/v0.0.1
 paths:
   /pins:
     get:
@@ -223,9 +247,5 @@ components:
       type: http
       scheme: bearer
       bearerFormat: JWT
-      # The JWT MUST contain a `user_id` claim that represents the user's identity within
-      # the backend system, so as to associate the user with their pins. This allows the 
-      # backend system to use the identity as a form of reference counting on pins, 
-      # therefore only removing pins once no more users have it pinned.
 security:
   - tokenAuth: []
